Add DistributedImpl and DistributedBackend arguments for micro benchmark.
